Added ortho camera controller, fixed some warnings and added functions for creating scopes and refs. Started working on a 2d example layer

+++ Engine/src/Renderer/Buffer.h
@@ -37,7 +37,7 @@ struct BufferElement
 {
 	std::string Name;
 	ShaderDataType Type;
-	uint32_t Offset;
+	size_t Offset;
 	uint32_t Size;
 	bool Normalized;
 
@@ -55,8 +55,8 @@ struct BufferElement
 		case ShaderDataType::Float2:	return 2;
 		case ShaderDataType::Float3:	return 3;
 		case ShaderDataType::Float4:	return 4;
-		case ShaderDataType::Mat3:		return 3*3;
-		case ShaderDataType::Mat4:		return 4*4;
+		case ShaderDataType::Mat3:		return 3 * 3;
+		case ShaderDataType::Mat4:		return 4 * 4;
 		case ShaderDataType::Int:		return 1;
 		case ShaderDataType::Int2:		return 2;
 		case ShaderDataType::Int3:		return 3;
@@ -90,7 +90,7 @@ public:
 private:
 	void CalculateOffsetsAndStride()
 	{
-		uint32_t offset = 0; 
+		size_t offset = 0; 
 		m_Stride = 0;
 		for (BufferElement& element : m_Elements)
 		{
